fix: cloneing objects
Did not really want to use a library for this but i have used for it to go away for now.

I think having the context and task to further debug errors are beneficial. Maybe we can make it
opt-in in the future.

But for having circular dependencies inside the context, this was causing problems without a deep
copy.

fix #571, fix #569, fix #538