perf: changes and reuses logger component and pluggable plugins for renderers
fix: #651, fix: #646, fix: #613

BREAKING CHANGE:
This completely breaks old behavior with the renderers, even the basic
parameters are changed.

Now the renderers for things like timer and timestamps takes in
pluggable components through the shared constants, which makes them user
configurable.

The logger interface and logger itself has changed to handle all the
common cases for any kind of renderer, eventhough it does not use all of
them at the same time.

This is a big change that might make the user checkout the new README.