feat: add task retry delay
fix: #668

Added delay functionality while trying to retry. Retry now can be an object that can also provide a
delay as well as the try count.
This is also done through the new PAUSED state for the task, the task will be paused whenever it is
retrying with delay now.