perf: removes options for disable/force colors on listr level
BREAKING CHANGE: `LISTR_DISABLE_COLOR` is dropped in favor of using `NO_COLOR` environment variable to make this not differenciate from other libraries.

Listr options `disableColor` and `forceColor` has been dropped since they were not working properly anyways.

The other way to do this was to make `color` in to a function that directly calls `createColors` from `colorrette`. But since it is used in constants and such it really did not make much sense.